requirements.txt
streamlit
pandas
supabase
graphviz
altair
python-dotenv
